rename t/smells-of-vcs/.svn to .exists and mkdir a .git instead
Commiting a file called .git to git would confuse things.

People (debian) who want to be able to import a tree into svn can't deal with
a .svn existing.

Therefore, we use .exists to ensure the directory's present, and then call
mkdir on t/smells-of-vcs/.git so the strictures behaviour still triggers.